Fixed bug in fn022 schema and added logic to fill in Fn028
This commit includes a fix to a bug in the FN022 schema that was being
caused by the inconsistence presence of timeszones on some dates.  It
also includes logic to pre-populate the FN028 table depending on the
selected gears. The list of gears available in the FN028 drop downs are
limited to those that were selected on the previous page.