Simplified form components by extracting errors in form
Originally the entire errors object was being passed to the form input
components and then extracted in the component. This required some
complicated logic to determine if the error was nested or not, and if
it was nested, what the index was.  This logic has been moved up to
the form instead where it is considerably simpler and the index is
available.  As a result, only errors associated with the field in
question are passed to the input component.