Added map navcontrol and some style to popup content.
This commit adds navcontrol components to both the click-able and point
maps, updated the marker with improved styling and placement, and add
some style to the popup label so that it isn't quite so bold. It also
removes several console.log statements that should have been removed
from earlier commits.